docs: generate with rustdocs
