+++ src/db/mod.rs
@@ -1,0 +1,11 @@
+/// Handles connections to multiple database engines
+pub mod conn;
+/// Handles database migrations
+///
+/// Note: All migrations are ran automatically when the application starts
+pub mod migration;
+/// Database models
+pub mod models;
+
+#[doc(hidden)]
+pub mod schema;
