+++ src/db/models.rs
@@ -3,6 +3,9 @@ use diesel::prelude::*;
 use serde::Serialize;
 use std::fmt::Display;
 
+use crate::utils::generate_uuid::generate_uuid;
+
+/// The status of a scan event being proccessed by [Targets](crate::service::targets).
 #[derive(Serialize)]
 pub enum ProcessStatus {
     Pending,
@@ -11,6 +14,9 @@ pub enum ProcessStatus {
     Failed,
 }
 
+/// Whether a file was found or not.
+///
+/// Note: only used if [opts.check_path](crate::utils::settings::Opts::check_path) is set.
 #[derive(Serialize)]
 pub enum FoundStatus {
     Found,
@@ -48,28 +54,45 @@ impl From<ProcessStatus> for String {
     }
 }
 
+/// Represents a scan event.
+///
+/// A scan event is created when a file is added by [Triggers](crate::service::triggers).
 #[derive(Queryable, Selectable, Serialize, Clone, Debug, AsChangeset, Identifiable)]
 #[diesel(table_name = crate::db::schema::scan_events)]
 pub struct ScanEvent {
+    /// The [uuid](crate::utils::generate_uuid::generate_uuid) of the scan event.
     pub id: String,
 
+    /// The name of the Trigger that created the scan event.
     pub event_source: String,
+    /// The time the scan event was created.
     pub event_timestamp: NaiveDateTime,
 
+    /// The rewritten path of the file.
     pub file_path: String,
+    /// Optional hash of the file.
     pub file_hash: Option<String>,
+    /// The status of the scan event being processed.
     pub process_status: String,
+    /// The status of the file being found.
     pub found_status: String,
 
+    /// The number of times the scan event has failed. Used for retries and is limited to [opts.max_retries](crate::utils::settings::Opts::max_retries).
     pub failed_times: i32,
+    /// The time the scan event will be retried.
     pub next_retry_at: Option<chrono::NaiveDateTime>,
 
+    /// The targets that have been hit by the scan event delimited by a comma.
     pub targets_hit: String,
 
+    /// The time the file was found.
     pub found_at: Option<chrono::NaiveDateTime>,
+    /// The time the scan event was processed.
     pub processed_at: Option<chrono::NaiveDateTime>,
 
+    /// The time the scan event was created.
     pub created_at: NaiveDateTime,
+    /// The time the scan event was updated.
     pub updated_at: NaiveDateTime,
 }
 
@@ -89,6 +112,7 @@ impl ScanEvent {
 
 #[derive(Insertable)]
 #[diesel(table_name = crate::db::schema::scan_events)]
+#[doc(hidden)]
 pub struct NewScanEvent {
     pub id: String,
     pub event_source: String,
@@ -97,11 +121,6 @@ pub struct NewScanEvent {
     pub file_hash: Option<String>,
 
     pub found_status: String,
-}
-
-fn generate_uuid() -> String {
-    let uuid = uuid::Uuid::new_v4();
-    uuid.to_string()
 }
 
 impl Default for NewScanEvent {
