+++ src/service/targets/command.rs
@@ -2,10 +2,22 @@ use crate::{db::models::ScanEvent, utils::settings::TargetProcess};
 use serde::Deserialize;
 use tracing::{debug, error};
 
+/// Command target
+///
+/// Note: Either `path` or `raw` must be set but not both
 #[derive(Clone, Deserialize)]
 pub struct Command {
+    /// Path to the command to run
+    ///
+    /// Example: `/path/to/script.sh`
     path: Option<String>,
+    /// Timeout for the command in seconds (default: 10)
+    ///
+    /// Example: `5`
     timeout: Option<u64>,
+    /// Raw command to run
+    ///
+    /// Example: `echo $FILE_PATH >> list.log`
     raw: Option<String>,
 }
 
@@ -20,9 +32,9 @@ impl Command {
                 .arg(&ev.file_path)
                 .output();
 
-            let timeout = self.timeout.unwrap_or(10000);
+            let timeout = self.timeout.unwrap_or(10);
 
-            let output = tokio::time::timeout(std::time::Duration::from_millis(timeout), output)
+            let output = tokio::time::timeout(std::time::Duration::from_secs(timeout), output)
                 .await
                 .map_err(|_| anyhow::anyhow!("command timed out"))??;
 
@@ -35,6 +47,7 @@ impl Command {
 
             debug!("command output: {:?}", output);
         }
+
         if let Some(raw) = self.raw.clone() {
             let output = tokio::process::Command::new("sh")
                 .env("FILE_PATH", &ev.file_path)
