+++ src/utils/generate_uuid.rs
@@ -1,0 +1,4 @@
+pub fn generate_uuid() -> String {
+    let uuid = uuid::Uuid::new_v4();
+    uuid.to_string()
+}
